+++ test_suite/tests/test_auto_enum.rs
@@ -9,6 +9,7 @@
         futures_api,
         generators,
         generator_trait,
+        iovec,
         read_initializer,
         trusted_len,
         try_trait,